new examples and folder names refactoring
