+++ pyinstaller/PythonTranslatorExcelAgMIP/PythonTranslatorExcelAgMIP.exe.manifest
