+++ pyinstaller/PythonTranslatorExcelAgMIP/base_library.zip
