+++ pyinstaller/PythonTranslatorExcelAgMIP/warn-PythonTranslatorExcelAgMIP.txt
